web layer impl

+++ src/main/java/topjava/grad/util/UserMock.java
@@ -3,6 +3,6 @@ package topjava.grad.util;
 import topjava.grad.domain.User;
 
 public class UserMock {
-    public static final Long USER_ID = 0L;
+    public static final Integer USER_ID = 0;
     public static final User USER = new User();
 }
